feat: Update NavSideMenu to conditionally render login button
The NavSideMenu component has been updated to conditionally render the login button based on the user's login status. If the user is logged in, the login button is replaced with a logout link. If the user is not logged in, the login button is displayed. This change improves the user experience by providing clear and intuitive navigation options.